Add menu creation manual test
We have found that there are three important operations
when it comes to menu creation and putting it up in the
native menubar:

    * Adding the action to its parent menu
    * Setting the submenu as the action's menu
    * Populating the actual submenu

This application tests all possible permutations for these
three operations, starting from the menubar and down two
more menu levels.

Two command-line options allow, first, to create a late
menubar that will replace the original menubar from the main
window form (--new-menubar) with, second, the option of
creating it parentless (--no-parent) before setting it as
the new main window menubar.

While mostly interesting for platforms supporting QPA menus
and menubars, this could be useful as a trivial check for
QMenu and QMenuBar in a pure QWidget environment.

As of today, the application shows at least three issues on
macOS which will be fixed in upcoming patches.

Change-Id: Id3c1e0f346c6fe27ab4656ff045b88a0a8623740
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>