Fix restoring geometry of dockwidget
In the past if the window with dock widgets was closed in maximized
mode, after reopening the dock widget size was different from the
original one. It was caused due to consecutive resizes of dock widgets
when the window was shown. During resizing dock widget geometry restored
from property file was lost. The solution was to  keep restored size of
dock widgets in QDockAreaLayoutInfo and use as hints for the dock layout
engine.

Task-number: QTBUG-16252
Change-Id: I94f4e895922063fa51150b68464fea4ddcb1d2a2
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>